# api/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Allinost_go-backend-core_internal_pkg_response.Response:
    properties:
      code:
        description: 业务错误码（0 表示成功）
        type: integer
      data:
        description: 响应数据
      detail:
        description: 错误详情（仅开发环境可见）
        type: string
      message:
        description: 提示信息
        type: string
      trace_id:
        description: 请求追踪 ID
        type: string
    type: object
...
  goodser.AllocateSeqReq:
    properties:
      inventory_id:
        type: string
      main_zone:
        type: string
      sub_zone:
        type: string
    type: object
  goodser.AllocateSeqResp:
    properties:
      seq_number:
        type: integer
    type: object
  goodser.CancelOutboundReq:
    properties:
      id:
        type: string
    type: object
  goodser.ConfirmOutboundReq:
    properties:
      id:
        type: string
    type: object
  goodser.CreateInventoryReq:
    properties:
      name:
        type: string
    type: object
  goodser.CreateOutboundReq:
    properties:
      inventory_id:
        type: string
      items:
        items:
          $ref: '#/definitions/goodser.OrderItem'
        type: array
      order_info:
        type: string
      order_no:
        type: string
      remark:
        type: string
      source_reserve_id:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  goodser.Inventory:
    properties:
      _id:
        type: string
      created_at:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  goodser.UpdateInventoryReq:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:29090
info:
  contact:
    name: Allinost
    url: https://github.com/Allinost/go-backend-core
  description: 模块化 Go 后端服务 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Allinost/go-backend-core
  title: Go Backend Core API
  version: 0.1.0
paths:
...
  /zzz-goodser/inventories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/goodser.Inventory'
                  type: array
              type: object
      summary: 获取目录列表
      tags:
      - zzz-goodser
    post:
      consumes:
      - application/json
      parameters:
      - description: 目录名
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CreateInventoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Inventory'
              type: object
      summary: 创建目录
      tags:
      - zzz-goodser
  /zzz-goodser/inventories/{id}:
    delete:
      parameters:
      - description: 目录 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 删除目录
      tags:
      - zzz-goodser
    get:
      parameters:
      - description: 目录 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Inventory'
              type: object
      summary: 获取目录
      tags:
      - zzz-goodser
    put:
      consumes:
      - application/json
      parameters:
      - description: 目录 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.UpdateInventoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Inventory'
              type: object
      summary: 更新目录
      tags:
      - zzz-goodser
  /zzz-goodser/inventories/{id}/products:
    get:
      parameters:
      - description: 库存目录 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/goodser.Product'
                  type: array
              type: object
      summary: 获取商品列表
      tags:
      - zzz-goodser
  /zzz-goodser/legacy/allocateSeq:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分配参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.AllocateSeqReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.AllocateSeqResp'
              type: object
      summary: 分配序号
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/cancelOutbound:
    post:
      consumes:
      - application/json
      description: 释放已锁定的预留库存
      parameters:
      - description: 出库单 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CancelOutboundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.OutboundOrder'
              type: object
      summary: 取消出库
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/confirmOutbound:
    post:
      consumes:
      - application/json
      description: 确认后扣除库存、取消预留
      parameters:
      - description: 出库单 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.ConfirmOutboundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.OutboundOrder'
              type: object
      summary: 确认出库
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/createInventory:
    post:
      consumes:
      - application/json
      parameters:
      - description: 目录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CreateInventoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Inventory'
              type: object
      summary: 创建目录
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/createOutbound:
    post:
      consumes:
      - application/json
      parameters:
      - description: 出库信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CreateOutboundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.OutboundOrder'
              type: object
      summary: 创建出库单
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/createProduct:
    post:
      consumes:
      - application/json
      description: 直接创建一个新商品，需先通过 allocateSeq 获取序号
      parameters:
      - description: 商品信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CreateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 创建商品
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/createTag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.CreateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Tag'
              type: object
      summary: 创建标签
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/deleteInventory:
    post:
      consumes:
      - application/json
      parameters:
      - description: 目录 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.DeleteInventoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 删除目录
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/deleteProduct:
    post:
      consumes:
      - application/json
      description: 删除指定商品并回收其序号
      parameters:
      - description: 商品 ID
        in: body
        name: body
        required: true
        schema:
          properties:
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 删除商品
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/inboundBatch:
    post:
      consumes:
      - application/json
      description: 批量创建或增加商品库存，已存在的按 code 匹配叠加数量
      parameters:
      - description: 入库项列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.InboundBatchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.InboundBatchResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 批量入库
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/inboundSearchImport:
    post:
      consumes:
      - application/json
      description: 将搜索结果中的商品批量导入（按 product_id 叠加库存）
      parameters:
      - description: 导入项列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.InboundSearchImportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.InboundSearchImportResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
      summary: 搜索导入入库
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/inboundSingle:
    post:
      consumes:
      - application/json
      description: 创 build 或叠加商品库存，写入库日志
      parameters:
      - description: 入库信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goodser.InboundSingleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/goodser.Product'
              type: object
      summary: 单品入库
      tags:
      - zzz-goodser-legacy
  /zzz-goodser/legacy/loadInboundLogs:
    post:
      consumes:
      - application/json
      parameters:
      - description: 仓库 ID
        in: body
        name: body
        required: true
        schema:
          properties:
            inventory_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Allinost_go-backend-core_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/goodser.InboundLog'
                  type: array
              type: object
      summary: 获取入库日志
      tags:
      - zzz-goodser-legacy
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token, 格式: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"